requirements.txt
streamlit==1.36.0
pandas==2.2.2
duckdb==0.10.3
python-dotenv==1.0.1

pydantic==2.7.4

langchain==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9 